feat: Bidirectional Inline Comment Navigation (#2068)
This adds a "reference" button to inline comment headers. This enables bidirectional scrolling behavior in that both the inline comments panel and the submission content area will scroll to make sure both the annotation and inline comment are visible whenever an inline comment's annotation or reference is clicked. Previously, this scrolling only happened within the inline comments panel when an annotation was clicked. Now, when an inline comment's reference is clicked, the corresponding annotation within the submission content area is scrolled to as well.

Closes #1936