ci: upgrade upload-artifact github action (#2142)
* Updates the usage of Ubuntu 20.04 to 24.04
* Updates the usage of upload-artifact from [v3 to v4](https://github.com/actions/upload-artifact/compare/v3...v4)
* Fixes a GraphQL-related timeout on `submissiondraft.cy.js`
* Adds hyphen delimiters for certain Vue attributes

I ran the tests three times for the latest passing commit so I'm reasonably confident that this is a reliable fix. However, there is a nondeterministic warning that appears for different attempts on the end-to-end testing: `Warning: No files were found with the provided path: ./client/test/cypress/screenshots. No artifacts will be uploaded.`